#39 否プリミティブ型に修正

+++ micro-api/src/main/java/com/mabubu0203/sudoku/api/controller/RestApiCreateController.java
@@ -69,7 +69,7 @@ public class RestApiCreateController extends RestBaseController {
             @PathVariable(name = "type") @Type final Integer type) {
 
         log.info("generateSudoku");
-        return service.generate(type);
+        return service.generate(type.intValue());
     }
 
 }
